서블릿 10일차 : 서블릿 복습 - 게시글 수정 예외처리 적용

+++ servlet/servlet_project2/src/main/java/kr/kh/app/service/BoardServiceImp.java
@@ -154,13 +154,15 @@ public class BoardServiceImp implements BoardService{
 			return false;
 		}
 		//삭제할 첨부파일 삭제
-		for(String numStr : nums) {
-			try {
-				int num = Integer.parseInt(numStr);
-				FileVO fileVo = boardDao.selectFile(num);
-				deleteFile(fileVo);
-			} catch (Exception e) {
-				e.printStackTrace();
+		if(nums != null) {
+			for(String numStr : nums) {
+				try {
+					int num = Integer.parseInt(numStr);
+					FileVO fileVo = boardDao.selectFile(num);
+					deleteFile(fileVo);
+				} catch (Exception e) {
+					e.printStackTrace();
+				}
 			}
 		}
